pyproject.toml
[tool.poetry]
name = "menprovning"
version = "0.1.0"
description = "AI-stött menprövningsverktyg för socialtjänsten"
authors = ["Menprövning Team"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
# PDF och dokumenthantering
pymupdf = "^1.24.0"
pytesseract = "^0.3.10"
pillow = "^10.1.0"
python-docx = "^1.1.0"

# NER och ML
transformers = "^4.36.0"
torch = "^2.1.0"
spacy = "^3.7.0"

# LLM
httpx = "^0.26.0"
tenacity = "^8.2.3"

# Data validation
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"

# API
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}

# Databas
sqlalchemy = "^2.0.25"
asyncpg = "^0.29.0"
alembic = "^1.13.0"

# Utilities
python-dateutil = "^2.8.2"
networkx = "^3.2.0"
rich = "^13.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.23.0"
pytest-benchmark = "^4.0.0"
ruff = "^0.1.11"
mypy = "^1.8.0"
pre-commit = "^3.6.0"
ipython = "^8.20.0"
types-python-dateutil = "^2.8.19"

[tool.poetry.group.test.dependencies]
faker = "^22.0.0"
hypothesis = "^6.92.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
    "validation: marks tests as validation tests",
]

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
